src/requirements.txt
# Core web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Database connections
psycopg2-binary>=2.9.7
qdrant-client>=1.6.0

# Machine Learning and NLP
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0

# Google Cloud Storage support
google-cloud-storage==2.10.0

# File processing
pdfplumber>=0.10.0
python-dotenv>=1.0.0

# LLM providers
openai>=1.3.0
google-generativeai>=0.3.0

# Cross-encoder for reranking
sentence-transformers

# Additional utilities
python-json-logger>=2.0.7
requests>=2.31.0
aiofiles>=23.2.0

# MinIO support (optional)
minio>=7.2.0
